requirements.txt
python-socketio==4.0.0
websockets==4.0.1
numpy==1.14.2
aiohttp==3.5.4
asn1crypto==0.24.0
async-timeout==3.0.1
certifi==2018.1.18
cffi==1.11.5
chardet==3.0.4
cryptography==2.1.4
Cython==0.28.2
gpg==1.10.0
HelperSDT==0.1.0
idna==2.6
idna-ssl==1.1.0
iniparse==0.4
multidict==3.3.2
ndg-httpsclient==0.4.4
nose==1.3.7
pandas==0.22.0
ply==3.11
pyalsaaudio==0.8.4
pyasn1==0.4.2
PyBluez==0.22
pycairo==1.15.6
pycparser==2.18
pygobject==3.28.1
pyOpenSSL==17.5.0
pyparsing==2.2.0
pyserial==3.4
PySocks==1.6.8
python-dateutil==2.7.2
python-engineio==3.5.0
python-periphery==1.1.1
pytz==2018.4
requests==2.18.4
six==1.11.0
smbus==1.1
spidev==3.2
urllib3==1.22
yarl==1.3.0